requirements.txt

pandas>=2.0.0
dearpygui>=1.11.1
matplotlib>=3.7.0
Pillow>=10.0.0
pyarrow>=16.0.0
numpy>=1.25.0
